requirements.txt
# Logging and CLI display
rich==14.1.0

# Token-aware chunking
tiktoken==0.9.0
regex==2024.5.15

# Embedding and Transformers
sentence-transformers==2.6.1
transformers>=4.32.0,<5.0.0
torch>=2.0.0,<3.0.0
faiss-cpu==1.7.4


# ML utilities
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.11.4

## Language models and tokenizers
transformers==4.54.0
accelerate==0.28.0
torch>=2.1.0
